% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{add_opt_col}
\alias{add_opt_col}
\title{add_opt_col}
\usage{
add_opt_col(lidar_trees, dfs = "all", qsm_idx_col = "qsm_idx",
  overwrite_existing_cols = F)
}
\arguments{
\item{lidar_trees}{concat_df[\link{"dataset}]}

\item{dfs}{dataframes to add columns to, Default: 'all'}

\item{qsm_idx_col}{name of qsm index column within dataframes, Default: 'qsm_idx'}

\item{overwrite_existing_cols}{Should existing columns in the dataframe be overwritten those from opt_models?, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
add "opt_set" and "opt_mod" columns to a dataframe, if the qsm_idx_col matches the opt_mod
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
