% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{add_tree_data}
\alias{add_tree_data}
\title{add_tree_data}
\usage{
add_tree_data(lidar_trees, dfs = "all", tree_data_df_name = "tree_data",
  data_cols = c("dbh", "pom", "height", "Astem_chambers_2004"),
  overwrite_existing_cols = F, match_col = "tree")
}
\arguments{
\item{lidar_trees}{PARAM_DESCRIPTION}

\item{dfs}{PARAM_DESCRIPTION, Default: 'all'}

\item{tree_data_df_name}{PARAM_DESCRIPTION, Default: 'tree_data'}

\item{data_cols}{PARAM_DESCRIPTION, Default: c("dbh", "pom", "height", "Astem_chambers_2004")}

\item{overwrite_existing_cols}{Should existing columns in the dataframe be overwritten those from tree_data?, Default: F}

\item{match_col}{PARAM_DESCRIPTION, Default: 'tree'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Adds data from the tree_data dataframe in a lidar_tree list/object to other dataframes in that list/object
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
