% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{assign_branch_num.BranchStructs}
\alias{assign_branch_num.BranchStructs}
\title{assign_branch_num.BranchStructs}
\usage{
\method{assign_branch_num}{BranchStructs}(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
assign branch numbers according to internodes occurring between furcations
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
