% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{assign_cyls_to_crown.TreeStructs}
\alias{assign_cyls_to_crown.TreeStructs}
\alias{assign_cyls_to_crown.default}
\title{assign_cyls_to_crown.TreeStructs}
\usage{
\method{assign_cyls_to_crown}{TreeStructs}(obj, daughter_threshold = 0.25)

\method{assign_cyls_to_crown}{default}(ts, first_branch_id)
}
\arguments{
\item{obj}{Treestructs object}

\item{daughter_threshold}{Proportion of the radius of the parent branch that daughter
branches must attain for a furction to be counted as the first branch, Default: 0.25}

\item{ts}{treestruct dataframe}
}
\value{
treestruct object with crown column assigned

treestruct dataframe
}
\description{
Assigns cylinders to crown and stem based on where the first branch occurs

Assigns logical crown column in a treestruct dataframe
}
\details{
Returns NA if no significant furcations found

Returns NA in crown column if no first branch was found
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
