% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{Astem_chambers_2004}
\alias{Astem_chambers_2004}
\title{Astem_chambers_2004}
\usage{
Astem_chambers_2004(DBH)
}
\arguments{
\item{DBH}{vector; tree diameter at breast height (cm)}
}
\value{
vector
}
\description{
Return surface area as estimate by Chambers' 2004 allometry
}
\details{
See Chambers 2004 and GEM manual
}
\examples{
Astem_chambers_2004(40)
}
