% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStruct.r
\name{BranchStruct}
\alias{BranchStruct}
\title{BranchStruct}
\usage{
BranchStruct(treestructdf = data.frame(), id = NA)
}
\arguments{
\item{treestructdf}{branch structure; see DESCRIPTION for column names, Default: data.frame()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Constructor for BranchStruct S3 object
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
