% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs.r
\name{BranchStructs}
\alias{BranchStructs}
\title{BranchStructs}
\usage{
BranchStructs(
  dataset = NA,
  treestructs,
  convert_to_meters = NA,
  has_topology = T
)
}
\arguments{
\item{dataset}{name of dataset this object is based on}

\item{treestructs}{dataframe of architecture data.  required.}

\item{surface_area_total}{total surface area of branch}

\item{pathlen_total}{total pathlength of branch (sum of branch lengths)}

\item{idcol}{name of branch id column; default = "branch_code"}

\item{datecol}{name of date column; default = "date"}

\item{internodeid_col}{name internode id column; default = "internode_id"}

\item{parentid_col}{name of parent internode id column; default = "parent_id"}

\item{furcation_col}{name of furcation number column; default = "n_furcation"}

\item{length_col}{name of internode length column; default = "len"}

\item{angle_col}{name of angle column; default = "major_child_angle"}

\item{angle_internode1_col}{name of angle internode 1 column; default = "angle_internode_1"}

\item{angle_internode2_col}{name of angle internode 2 column; default = "angle_internode_2"}

\item{numnodes_col}{name of number of nodes in section column; default = "number_nodes_in_section"}

\item{istip_col}{name of is tip column; default = "is_tip"}

\item{isbroken_col}{name of is broken column; default = "is_broken"}

\item{d_child_col}{name of internode diameter @ child (top) column; default = "d_child"}

\item{d_parent_col}{name of internode diameter @ parent (base) column; default = "d_parent"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
BranchStructs are hand-measured branches or trees.  This is a constructor for the BranchStructs S3 object.  Set column names before importing the treestruct dataframe
}
\details{
To create a branchstruct, pass a dataset with the following columns (unless you specify different column names, which is not recommended):
branch_code
date
internode_id
parent_id
n_furcation (optional)
len
major_child_angle (optional)
angle_internode_1 (optional)
angle_internode_2 (optional)
number_nodes_in_section (optional)
is_tip
is_broken
d_child
d_parent
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
