% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStruct_methods.r
\name{calc_pathlen.BranchStruct}
\alias{calc_pathlen.BranchStruct}
\title{calc_pathlen.BranchStruct}
\usage{
\method{calc_pathlen}{BranchStruct}(obj)
}
\arguments{
\item{obj}{BranchStruct object}
}
\value{
BranchStruct object
}
\description{
calc path lengths
}
\details{
overwrites treestruct::calc_pathlen
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
