% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{calc_pathlen.BranchStructs}
\alias{calc_pathlen.BranchStructs}
\title{calc_pathlen.BranchStructs}
\usage{
\method{calc_pathlen}{BranchStructs}(bss)
}
\arguments{
\item{obj}{BranchStructs object}
}
\value{
BranchStructs object
}
\description{
calc path lengths
}
\details{
Details
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
