% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{calc_pathlen.default}
\alias{calc_pathlen.default}
\title{calc_pathlen.default}
\usage{
\method{calc_pathlen}{default}(tree_structure, length_col = "len",
  parent_row_col = "parent_row", path_len_col = "path_len")
}
\arguments{
\item{tree_structure}{a tree structure dataframe}

\item{length_col}{chr specifiying length column. default:"len"}

\item{parent_row_col}{chr specifiying parent row column. default:"parent_row"}

\item{path_len_col}{chr specifiying path length column. default:"path_len"}
}
\value{
tree structure dataframe with a path_len column populated
}
\description{
Wrapper for C++ pathlength code
}
\details{
DETAILS
}
