% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{calc_pathlen}
\alias{calc_pathlen}
\title{calc_pathlen}
\usage{
calc_pathlen(tree_structure)
}
\arguments{
\item{tree_structure}{a tree structure dataframe}
}
\value{
tree structure dataframe with a path_len column populated
}
\description{
Wrapper for C++ pathlength code
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
