% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{calc_pathlen.TreeStructs}
\alias{calc_pathlen.TreeStructs}
\title{calc_pathlen.TreeStructs}
\usage{
\method{calc_pathlen}{TreeStructs}(obj)
}
\arguments{
\item{obj}{TreeStructs object}
}
\value{
TreeStructs object
}
\description{
calc path lengths
}
\details{
Details
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
