% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{calc_sa_above.BranchStructs}
\alias{calc_sa_above.BranchStructs}
\title{Calculate surface area above every node}
\usage{
\method{calc_sa_above}{BranchStructs}(obj)
}
\arguments{
\item{obj}{Branchstruct object}
}
\value{
Branchstruct object
}
\description{
Calculate surface area above every node
}
\details{
an 'sa_above' column is added to the treestruct nested dataframe
}
