% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{calc_sa_above}
\alias{calc_sa_above}
\title{FUNCTION_TITLE}
\usage{
calc_sa_above(tree_structure)
}
\arguments{
\item{tree_structure}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
