% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{calc_sa_above.TreeStructs}
\alias{calc_sa_above.TreeStructs}
\title{Calculate surface area above every node}
\usage{
\method{calc_sa_above}{TreeStructs}(obj)
}
\arguments{
\item{obj}{Treestruct object}
}
\value{
treestruct object
}
\description{
Calculate surface area above every node
}
\details{
an 'sa_above' column is added to the cyl_summ nested dataframe
}
