% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStruct_methods.r
\name{calc_surfarea}
\alias{calc_surfarea}
\title{FUNCTION_TITLE}
\usage{
calc_surfarea(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
