% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{calc_surfarea.TreeStructs}
\alias{calc_surfarea.TreeStructs}
\title{FUNCTION_TITLE}
\usage{
\method{calc_surfarea}{TreeStructs}(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
