% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{calc_vol_above.BranchStructs}
\alias{calc_vol_above.BranchStructs}
\title{Calculate wood volume above every node}
\usage{
\method{calc_vol_above}{BranchStructs}(obj)
}
\arguments{
\item{obj}{Branchstruct object}
}
\value{
Branchstruct object
}
\description{
Calculate wood volume above every node
}
\details{
a 'vol_above' column is added to the treestruct nested dataframe
}
