% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{calc_vol_above.TreeStructs}
\alias{calc_vol_above.TreeStructs}
\title{Calculate wood volume above every node}
\usage{
\method{calc_vol_above}{TreeStructs}(obj)
}
\arguments{
\item{obj}{Treestruct object}
}
\value{
treestruct object
}
\description{
Calculate wood volume above every node
}
\details{
a 'vol_above' column is added to the cyl_summ nested dataframe
}
