% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{concat_lidar_trees_parallel}
\alias{concat_lidar_trees_parallel}
\title{FUNCTION_TITLE}
\usage{
concat_lidar_trees_parallel(lidar_trees, pick_qsm = F, best_qsm = "")
}
\arguments{
\item{lidar_trees}{PARAM_DESCRIPTION}

\item{pick_qsm}{PARAM_DESCRIPTION, Default: F}

\item{best_qsm}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link{concat_lidar_trees}}
}
