% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{filter_opt_mods_df}
\alias{filter_opt_mods_df}
\title{filter_opt_mods_df}
\usage{
filter_opt_mods_df(df, qsm_idx_col = "qsm_idx", opt_qsm_idxs)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{qsm_idx_col}{PARAM_DESCRIPTION, Default: 'qsm_idx'}

\item{opt_qsm_idxs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
keep only optimal models in a data.frame
}
\details{
If opt_qsm_idxs is empty, then the full dataframe is return unaltered
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
