% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{find_first_branch.TreeStructs}
\alias{find_first_branch.TreeStructs}
\alias{find_first_branch.default}
\title{find_first_branch.TreeStructs}
\usage{
\method{find_first_branch}{TreeStructs}(obj, daughter_threshold = 0.25)

\method{find_first_branch}{default}(ts, daughter_threshold = 0.25)
}
\arguments{
\item{obj}{Treestructs object}

\item{daughter_threshold}{Proportion of the radius of the parent branch that daughter
branches must attain for a furction to be counted as the first branch, Default: 0.25}

\item{ts}{treestruct dataframe}
}
\value{
internode_id of parent branch where first furcation occurs

internode_id of parent branch where first furcation occurs
}
\description{
Finds the first branches of treestructs

Finds the first branch in a treestruct dataframe
}
\details{
Returns NA if no significant furcations found

Returns NA if no significant furcations found
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
