% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{get_all_branches}
\alias{get_all_branches}
\title{Get all branches of a specified basal radius from a treestruct dataframe}
\usage{
get_all_branches(ts, target_rad, range)
}
\arguments{
\item{ts}{treestruct dataframe}

\item{target_rad}{target basal radius (in m)}

\item{range}{what range around the target radius will you look for branches?}
}
\value{
list of treestruct data frames that are the branches that fit the criteria
}
\description{
Get all branches of a specified basal radius from a treestruct dataframe
}
\details{

}
