% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{get_child_rows}
\alias{get_child_rows}
\title{Get children of internode}
\usage{
get_child_rows(ts, internode_id, row)
}
\arguments{
\item{ts}{treestruct dataframe (not object)}

\item{internode_id}{if this is provided, the id of the internode for which children are sought}

\item{row}{if this is provided, the row number of the internode for which children are sought}
}
\value{
vector of rows of internode children, including the row of the specified internode
}
\description{
Get children of internode
}
\details{
only one of internode_id and row arguments are required.
}
