% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{getFile}
\alias{getFile}
\title{getFile}
\usage{
getFile(obj)
}
\arguments{
\item{obj}{TreeStructs or BranchStructs object}
}
\value{
vector of base file names
}
\description{
file list accessor
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
