% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{list_cols}
\alias{list_cols}
\title{FUNCTION_TITLE}
\usage{
list_cols(df, return_names = TRUE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{return_names}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
