% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{parent_row}
\alias{parent_row}
\title{Calculate the parent_row column}
\usage{
parent_row(parent_id, internode_id)
}
\arguments{
\item{internode_id}{}
}
\description{
Calculate the parent_row column
}
