% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{parse_id.BranchStructs}
\alias{parse_id.BranchStructs}
\title{parse_id.BranchStructs}
\usage{
\method{parse_id}{BranchStructs}(
  obj,
  treetag_regex = ".*",
  branch_regex = "B\\\\d+(SH|S|H)",
  treetag_sec_no = 2,
  branch_sec_no = 3,
  nobranchcode = F,
  ...
)
}
\arguments{
\item{obj}{treestruct object}

\item{treetag_regex}{regex to match tree tag, Default: \verb{.*}}

\item{branch_regex}{regex to match branchcode, Default: \verb{B\\\\d+(SH|S|H)}}

\item{treetag_sec_no}{which section of the string (separated by \verb{[_-]}) contains the treetag? Default: 2}

\item{branch_sec_no}{which section of the string (separated by \verb{[_-]}) contains the branch code? Default: 3}

\item{nobranchcode}{set to TRUE if there is no branch code in the identifier, Default: FALSE}
}
\value{
treestruct object
}
\description{
parse the plot, tree tag, and branch identifier out from the concatenated identifier
}
\details{
This function assumes the id column (idcol) is in the form of "plotcode-treecode-branchcode"
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
