% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{process_qsm_dir}
\alias{process_qsm_dir}
\title{FUNCTION_TITLE}
\usage{
process_qsm_dir(qsm_path = ".", parallel_process = T,
  cyl_file_pat = "cyl.*.txt", recursive = F, file_batching = 0)
}
\arguments{
\item{qsm_path}{PARAM_DESCRIPTION}

\item{parallel_process}{PARAM_DESCRIPTION, Default: T}

\item{cyl_file_pat}{PARAM_DESCRIPTION, Default: 'cyl.*.txt'}

\item{recursive}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
