% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{readQSM.mat}
\alias{readQSM.mat}
\title{readQSM.mat}
\usage{
readQSM.mat(qsmfile, qsmver = 2.3)
}
\arguments{
\item{qsmfile}{path to .mat file}

\item{qsmver}{version of treeQSM used to make the .mat file, Default: 2.3}
}
\value{
a 3-element list of file (filename), CylData, and BranchData
}
\description{
reads in a QSM file in matlab format and passes back cylinder and branch data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
