% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{readQSM}
\alias{readQSM}
\title{readQSM}
\usage{
readQSM(
  qsmfile,
  qsmtype = c("by_ext", "QSMtree", "treegraph"),
  qsmver = "by_name"
)
}
\arguments{
\item{qsmfile}{path to file}

\item{qsmtype}{program used to create QSM - currently either QSMtree (.mat) or treegraph (.json). Default: "by_ext" (i.e., .mat = QSMtree, .json = treegraph)}

\item{qsmver}{version of treeQSM used to make the .mat file, Default: "by_name"}
}
\value{
a 3-element list of file (filename), CylData, and BranchData
}
\description{
reads in a QSM file by dispatching to readQSM.mat or readQSM.treegraph
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
