% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{readQSM.treegraph}
\alias{readQSM.treegraph}
\title{readQSM.treegraph}
\usage{
readQSM.treegraph(qsmfile, qsmver = "by_name")
}
\arguments{
\item{qsmfile}{path to treegraph file}

\item{qsmver}{version of treegraph used to make the treegraph file, Default: by_name}
}
\value{
a 3-element list of file (filename), CylData, and BranchData
}
\description{
reads in a QSM file in treegraph format and passes back cylinder and branch data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
