% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{rename_tree}
\alias{rename_tree}
\title{Rename lidar tree}
\usage{
rename_tree(name, regex, name_is_path = T)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}

\item{regex}{PARAM_DESCRIPTION}

\item{name_is_path}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
convenience function to name a QSM tree based on it's file name
}
\details{
DETAILS
}
\note{
This is only exported becuase it's used internally in foreach
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
