% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{reorder_internodes.BranchStructs}
\alias{reorder_internodes.BranchStructs}
\title{reorder_internodes.BranchStructs}
\usage{
\method{reorder_internodes}{BranchStructs}(obj)
}
\arguments{
\item{obj}{TreeStructs object}
}
\value{
TreeStructs object
}
\description{
reorders internodes in treestruct objects so pathlength and other accumulation routines work properly
}
\details{
This reorders all rows in the obj$treestructs$treestruct dataframes. It runs setTips if that hasn't already
been run.  This reorders rows such that, as you go down the dataframe, you can accumulate metrics (such as pathlength and
surface area).  That is, you are guaranteed that internodes above you (children, grandchildren, etc) will have been
assigned values.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
