% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{sapwood_volume}
\alias{sapwood_volume}
\title{FUNCTION_TITLE}
\usage{
sapwood_volume(lidar_tree, sapwood_depth = 2)
}
\arguments{
\item{lidar_tree}{PARAM_DESCRIPTION}

\item{sapwood_depth}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
