% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{save_cylfiles}
\alias{save_cylfiles}
\title{save_cylfiles}
\usage{
save_cylfiles(obj, ...)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Extract cylinder files from treestruct objects and save them to csv files.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
