% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{save_cylfiles.TreeStructs}
\alias{save_cylfiles.TreeStructs}
\title{FUNCTION_TITLE}
\usage{
\method{save_cylfiles}{TreeStructs}(obj, path = ".")
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
