% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{save_model_text_matlab_server}
\alias{save_model_text_matlab_server}
\title{save_model_text_matlab_server}
\usage{
save_model_text_matlab_server(qsm_path = ".", recursive = F,
  treeqsm_path = "")
}
\arguments{
\item{qsm_path}{PARAM_DESCRIPTION, Default: '.'}

\item{recursive}{PARAM_DESCRIPTION, Default: F}

\item{treeqsm_path}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Experimental and not tested!  Uses connection to matlab server to process .mat QSM files.
}
\details{
this may not work if your version of treeqsm doesn't match the version used when making the .mat files.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
