% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{set_valid_col}
\alias{set_valid_col}
\title{set_valid_col}
\usage{
set_valid_col(obj)
}
\arguments{
\item{obj}{TreeStruct or BranchStruct object}
}
\value{
TreeStruct or BranchStruct object
}
\description{
set overall validity property of treestruct/branchstruct members
}
\details{
Removes 'valid' column if it already exists, so we don't include it in determining whether a treestruct is valid,
and then looks for all columns with "valid" in their name.  Performs a logical 'and' on those columns in a new column, 'valid'.
}
