% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{setDataset.default}
\alias{setDataset.default}
\title{FUNCTION_TITLE}
\usage{
\method{setDataset}{default}(obj, newVal)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{newVal}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
