% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{setGraph.BranchStructs}
\alias{setGraph.BranchStructs}
\title{setGraph.BranchStructs}
\usage{
\method{setGraph}{BranchStructs}(obj, ts_accessor = getTreestruct)
}
\arguments{
\item{obj}{BranchStructs object}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Invalid treestructs will have graph set to NA.  This means that all functions making use of graphs will have to
be able to handle NA as an input in addition to a graph object.  This is a design decision.  An alternative is to use an empty graph object, but then
it becomes less obvious which treestructs are invalid and hence don't have graphs, and could lead to including erroneous
data in analyses.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
