% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{setGraph.TreeStructs}
\alias{setGraph.TreeStructs}
\title{setGraph.TreeStructs}
\usage{
\method{setGraph}{TreeStructs}(obj)
}
\arguments{
\item{obj}{TreeStructs object}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
