% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStruct_methods.r,
%   R/BranchStructs_methods.r
\name{setID}
\alias{setID}
\alias{setDataset}
\title{FUNCTION_TITLE}
\usage{
setID(obj, newVal)

setDataset(obj, newVal)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{newVal}{PARAM_DESCRIPTION}

\item{obj}{PARAM_DESCRIPTION}

\item{newVal}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
