% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStruct_methods.r
\name{setTreestruct.BranchStruct}
\alias{setTreestruct.BranchStruct}
\title{FUNCTION_TITLE}
\usage{
\method{setTreestruct}{BranchStruct}(obj, treestruct)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{treestruct}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
