% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{setTreestruct.BranchStructs}
\alias{setTreestruct.BranchStructs}
\title{setTreestruct.BranchStructs}
\usage{
\method{setTreestruct}{BranchStructs}(obj, treestructs)
}
\arguments{
\item{obj}{BranchStructs object}

\item{treestructs}{treestruct data frame}
}
\value{
BranchStructs object with validated, nested, treestructs dataframe.
}
\description{
This validates and creates the nested dataframe and populates the object with it.
}
\details{
We use the column names defined in the object properties.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
