% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{setTreestruct.TreeStructs}
\alias{setTreestruct.TreeStructs}
\title{setTreestruct.TreeStructs}
\usage{
\method{setTreestruct}{TreeStructs}(obj, treestructs)
}
\arguments{
\item{obj}{TreeStructs object}

\item{treestructs}{treestruct data frame}
}
\value{
TreeStructs object with validated, nested, treestructs dataframe.
}
\description{
This validates and creates the nested dataframe and populates the object with it.
}
\details{
We use the column names defined in the object properties.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
