% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{truncate_branches}
\alias{truncate_branches}
\title{truncate_branches}
\usage{
truncate_branches(obj, ...)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{rad_min}{minimum radius of branch to keep in meters.  0.05 would mean keeping all branches >= 5cm.}
}
\value{

}
\description{
removes branches smaller than a given radius
}
\details{
make sure to re-run routines that calculate tree metrics after altering the tree
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
