% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{validate_internode_order}
\alias{validate_internode_order}
\title{validate_internode_order}
\usage{
validate_internode_order(parents, internode_id = NA, parents_are_rows = T)
}
\arguments{
\item{parents}{vector of parent internodes, either rows or id's (if id's, then internode_id parameter required)}

\item{internode_id}{vector of internode ids, only required if parents are ids, not rows, default: NA}

\item{parents_are_rows}{boolean flag to indicate whether parent vector refers to parent rows or ids, default: T}
}
\value{
boolean
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
