% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{validate_parents}
\alias{validate_parents}
\title{validate_parents}
\usage{
validate_parents(
  internode_ids,
  parent_ids,
  ignore_errors = NA,
  parents_are_rows = F,
  ignore_base_id = T
)
}
\arguments{
\item{internode_ids}{vector}

\item{parent_ids}{vector}

\item{ignore_errors}{Boolean vector specifying which rows to ignore errors on.  A value of NA means no erros will be ignored.}

\item{parents_are_rows}{If parent_ids are actually row numbers, not id's (common for QSM cyl_files), Default: F}
}
\value{
TRUE if validation passes, FALSE if it fails
}
\description{
check whether the parent/child structure tree structure data is valid (i.e., does every parent refer to an existing internode)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
