% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{visNetwork.BranchStructs}
\alias{visNetwork.BranchStructs}
\title{visNetwork.BranchStructs}
\usage{
\method{visNetwork}{BranchStructs}(bss, index, hierarchical = T,
  width_factor = 100, length_factor = 10)
}
\arguments{
\item{bss}{BranchStructs object (requred)}

\item{index}{\code{character} or \code{integer} char matching BranchStructs id column, or index of branch in BranchStructs object (required)}

\item{hierarchical}{\code{logical}, plot in hierarchical layout. Default: T}

\item{width_factor}{\code{numeric} edge width factor, Default: 100}

\item{length_factor}{\code{numeric} edge length factor, Default: 10}
}
\value{
visNetwork object
}
\description{
visualize branch network.  This overwrites visNetwork for BranchStructs objects.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 visNetwork(bss_obj) \%>\% visEdges(arrows = "middle")
 }
}
}
\seealso{
\code{\link[visNetwork]{visNetwork-igraph}},\code{\link[visNetwork]{visNetwork}}
}
