% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_funcs.r
\name{write_files_from_mats_dir}
\alias{write_files_from_mats_dir}
\title{write_files_from_mats_dir}
\usage{
write_files_from_mats_dir(qsm_path = ".", recursive = F,
  qsmver = 2.3)
}
\arguments{
\item{qsm_path}{path to directory containing treeQSM .mat files}

\item{recursive}{process sub-directories recursively?  Default: FALSE}

\item{qsmver}{version of treeQSM used to make the .mat file, Default: 2.3}
}
\value{
vector of .mat filenames processed
}
\description{
extracts cyl and branch data from .mat files made by treeQSM and writes txt files used by this package.
}
\details{
If you get an error about names not being the same length as the vector, try a different qsm version.  This is probably slower than the equivalent server function, but also more reliable.  This functions does not use matlab at all.  Warning: this will overwrite existing files.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
