% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{analyze_cyl_file}
\alias{analyze_cyl_file}
\title{FUNCTION_TITLE}
\usage{
analyze_cyl_file(cyl_file, calc_sapwood = T, sapwood_depth = 2, treeid)
}
\arguments{
\item{cyl_file}{PARAM_DESCRIPTION}

\item{calc_sapwood}{PARAM_DESCRIPTION, Default: T}

\item{sapwood_depth}{PARAM_DESCRIPTION, Default: 2}

\item{treeid}{identifier to use in outputs for this tree, Default: basename(cyl_file)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
