% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs.r
\name{BranchStructs}
\alias{BranchStructs}
\title{BranchStructs}
\usage{
BranchStructs(dataset = NA, treestructs)
}
\arguments{
\item{dataset}{name of dataset this object is based on}

\item{treestructs}{dataframe of architecture data.  required.}

\item{idcol}{name of branch id column; default = "branch_code"}

\item{datecol}{name of date column; default = "date"}

\item{internodeid_col}{name internode id column; default = "internode_id"}

\item{parentid_col}{name of parent internode id column; default = "parent_id"}

\item{furcation_col}{name of furcation number column; default = "n_furcation"}

\item{length_col}{name of internode length column; default = "len"}

\item{angle_col}{name of angle column; default = "major_child_angle"}

\item{angle_internode1_col}{name of angle internode 1 column; default = "angle_internode_1"}

\item{angle_internode2_col}{name of angle internode 2 column; default = "angle_internode_2"}

\item{numnodes_col}{name of number of nodes in section column; default = "number_nodes_in_section"}

\item{istip_col}{name of is tip column; default = "is_tip"}

\item{isbroken_col}{name of is broken column; default = "is_broken"}

\item{d_child_col}{name of internode diameter @ child (top) column; default = "d_child"}

\item{d_parent_col}{name of internode diameter @ parent (base) column; default = "d_parent"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Constructor for BranchStructs S3 object.  Set column names before importing the treestruct dataframe
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
