% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{parse_id.BranchStructs}
\alias{parse_id.BranchStructs}
\title{parse_id.BranchStructs}
\usage{
\method{parse_id}{BranchStructs}(
  obj,
  treetag_regex = ".*",
  branch_regex = "B\\\\d+[S][H]?",
  nobranchcode = F,
  ...
)
}
\arguments{
\item{obj}{treestruct object}

\item{treetag_regex}{regex to match tree tag, Default: '.*'}

\item{branch_regex}{regex to match branchcode, Default: 'B\\ded+\link[=H]{S}?'}

\item{nobranchcode}{set to TRUE if there is no branch code in the identifier, Default: FALSE}
}
\value{
treestruct object
}
\description{
parse the plot, tree tag, and branch identifier out from the concatenated identifier
}
\details{
This function assumes the id column (idcol) is in the form of "plotcode-treecode-branchcode"
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
