% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{pathlengths}
\alias{pathlengths}
\title{pathlengths}
\usage{
pathlengths(treestruct, start_order)
}
\arguments{
\item{treestruct}{PARAM_DESCRIPTION}

\item{start_order}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
