% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_funcs.r
\name{process_qsm_dir}
\alias{process_qsm_dir}
\title{process_qsm_dir}
\usage{
process_qsm_dir(qsm_path = ".", parallel_process = T,
  cyl_file_pat = "cyl.*.txt", rename_pat = "", recursive = F,
  file_batching = 0)
}
\arguments{
\item{qsm_path}{PARAM_DESCRIPTION}

\item{parallel_process}{PARAM_DESCRIPTION, Default: T}

\item{cyl_file_pat}{PARAM_DESCRIPTION, Default: 'cyl.*.txt'}

\item{rename_pat}{Regex applied to filenames when naming ouput list elements (sub(rename_pat, '\1', filename))}

\item{recursive}{PARAM_DESCRIPTION, Default: F}

\item{file_batching}{Number of files to process per batch, or 0 to process all at once, Default: 0}
}
\value{
list, one element per cyl file, of outputs from analyze_cyl_file
}
\description{
Runs analyze_cyl_file on all files in a directory
}
\details{
This routine provides a facility to name each element of the list based on the name of the corresponding cyl file.  Use rename_pat parameter to customize the renaming.  File batching can be used to avoid memory errors, etc
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
