% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{setGraph.BranchStructs}
\alias{setGraph.BranchStructs}
\title{setGraph.BranchStructs}
\usage{
\method{setGraph}{BranchStructs}(obj, ts_accessor = getTreestruct)
}
\arguments{
\item{obj}{BranchStructs object}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
