% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchStructs_methods.r
\name{setTreestruct.BranchStructs}
\alias{setTreestruct.BranchStructs}
\title{setTreestruct.BranchStructs}
\usage{
\method{setTreestruct}{BranchStructs}(obj, treestructs, convert_to_meters = T, abort_when_invalid = F)
}
\arguments{
\item{obj}{BranchStructs object}

\item{treestructs}{treestruct data frame}

\item{convert_to_meters}{Hand-measured branches typically use mm for diameter and cm for length.  treestruct uses meters as a standard unit.}
}
\value{
BranchStructs object with validated, nested, treestructs dataframe.
}
\description{
This validates and creates the nested dataframe and populates the object with it.
}
\details{
We use the column names defined in the object properties.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
