% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStructs_methods.r
\name{setTreestruct.TreeStructs}
\alias{setTreestruct.TreeStructs}
\title{setTreestruct.TreeStructs}
\usage{
\method{setTreestruct}{TreeStructs}(obj, treestructs, convert_to_meters = NA, abort_when_invalid = F)
}
\arguments{
\item{obj}{TreeStructs object}

\item{treestructs}{treestruct data frame}

\item{convert_to_meters}{numeric.  TreeQSM and treegraph use meters as a standard unit.  If units are not in meters, pass the conversion factor here such that meters = QSMunits * convert_to_meters.  Only radius and length values are converted, not cylinder positions or axes.  A value of NA results in no conversion.  Default: NA}
}
\value{
TreeStructs object with validated, nested, treestructs dataframe.
}
\description{
This validates and creates the nested dataframe and populates the object with it.
}
\details{
We use the column names defined in the object properties.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
